requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
et-xmlfile==1.1.0
filelock==3.10.7
idna==3.4
openpyxl==3.1.2
python-dateutil==2.8.2
requests==2.28.2
requests-file==1.5.1
six==1.16.0
tldextract==3.4.0
urllib3==1.26.15
